Refactor Claim
