+++ app/modules/claim.js
@@ -1,5 +1,9 @@
 // @flow
-import { getClaimAmounts } from 'neon-js'
+import { doClaimAllGas, doSendAsset, getClaimAmounts } from 'neon-js'
+import { sendEvent, clearTransactionEvent } from '../modules/transactions'
+import { log } from '../util/Logs'
+import { ASSETS } from '../core/constants'
+
 // Constants
 export const SET_CLAIM = 'SET_CLAIM'
 export const SET_CLAIM_REQUEST = 'SET_CLAIM_REQUEST'
@@ -32,6 +36,40 @@ export const syncAvailableClaim = (net: NetworkType, address: string) => (dispat
   getClaimAmounts(net, address).then((result) => {
     return dispatch(setClaim(result.available, result.unavailable))
   })
+}
+
+export const doClaimNotify = (net: NetworkType, wif: string, address: string) => (dispatch: DispatchType) => {
+  log(net, 'CLAIM', address, { info: 'claim all gas' })
+  doClaimAllGas(net, wif).then((response) => {
+    if (response.result) {
+      dispatch(sendEvent(true, 'Claim was successful! Your balance will update once the blockchain has processed it.'))
+      setTimeout(() => dispatch(disableClaim(false)), 300000)
+    } else {
+      dispatch(sendEvent(false, 'Claim failed'))
+    }
+    setTimeout(() => dispatch(clearTransactionEvent()), 5000)
+  })
+}
+
+// To initiate claim, first send all Neo to own address, the set claimRequest state
+// When new claims are available, this will trigger the claim
+export const doGasClaim = (net: NetworkType, wif: string, address: string, neo: number) => (dispatch: DispatchType) => {
+  // if no neo in account, no need to send to self first
+  if (neo === 0) {
+    dispatch(doClaimNotify(net, wif, address))
+  } else {
+    dispatch(sendEvent(true, 'Sending Neo to Yourself...'))
+    log(net, 'SEND', address, { to: address, amount: neo, asset: 'NEO' })
+    doSendAsset(net, address, wif, { [ASSETS.NEO]: neo }).then((response) => {
+      if (response.result === undefined || response.result === false) {
+        dispatch(sendEvent(false, 'Transaction failed!'))
+      } else {
+        dispatch(sendEvent(true, 'Waiting for transaction to clear...'))
+        dispatch(setClaimRequest(true))
+        dispatch(disableClaim(true))
+      }
+    })
+  }
 }
 
 // Reducer for managing claims data
